// src/interfaces/monitor.ts
export interface IMonitorDocument {
    id?: number | string;
    monitorId?: number;
    notificationId?: number;
    name: string;
    active?: boolean;
    status?: number;
    userId: number;
    frequency: number;
    alertThreshold: number | string;
    url?: string;
    type: string;
    lastChanged?: Date | string;
    timeout?: number | string;
    uptime?: number;
    redirects?: number | string;
    method?: string;
    headers?: string;
    body?: string;
    httpAuthMethod?: string;
    basicAuthUser?: string;
    basicAuthPass?: string;
    bearerToken?: string;
    contentType?: string;
    statusCode?: string;
    responseTime?: string | number;
    connection?: string;
    port?: number | string;
    heartbeats?: IHeartbeat[];
    notifications?: INotification;
}

export interface IMonitorErrorMessage {
    [x: string]: string | undefined;
    notificationId?: string;
    name?: string;
    frequency?: string;
    alertThreshold?: string;
    url?: string;
    timeout?: string;
    redirects?: string;
    method?: string;
    headers?: string;
    body?: string;
    httpAuthMethod?: string;
    basicAuthUser?: string;
    basicAuthPass?: string;
    bearerToken?: string;
    contentType?: string;
    statusCode?: string;
    responseTime?: string;
    connection?: string;
    port?: string;
}
// ...
export interface IHeartbeat {
    id?: number;
    monitorId: number;
    status: number;
    code: number;
    message: string;
    timestamp: number | string;
    reqHeaders?: string;
    resHeaders?: string;
    reqBody?: string;
    resBody?: string;
    responseTime: number;
    connection?: string;
}